A few comment weaks; move execShell
